renamed remaining controller files to coordinator. tweaked build to use a different virtual environment folder so as not to conflict with PyCharm.
