+++ open-tofu/main.tf.template
@@ -51,13 +51,13 @@ resource "aws_iam_role_policy_attachment" "invoke-lambda-policy-attach" {
 }
 
 
-# Init the lambda controller.
-resource "aws_lambda_function" "lambda_controller" {
-    filename      = "../{{source-path}}/lambda_controller/lambda_controller.zip"
-    function_name = "open_mpic_lambda_controller_{{deployment-id}}"
+# Init the lambda mpic coordinator.
+resource "aws_lambda_function" "lambda_mpic_coordinator" {
+    filename      = "../{{source-path}}/lambda_mpic_coordinator/lambda_mpic_coordinator.zip"
+    function_name = "open_mpic_lambda_coordinator_{{deployment-id}}"
     role          = aws_iam_role.open_mpic_lambda_role.arn
     handler       = "mpic_coordinator_lambda_function.lambda_handler"
-    source_code_hash = filebase64sha256("../{{source-path}}/lambda_controller/lambda_controller.zip")
+    source_code_hash = filebase64sha256("../{{source-path}}/lambda_mpic_coordinator/lambda_mpic_coordinator.zip")
     runtime = "python3.11"
     timeout = 60
     layers = [aws_lambda_layer_version.python311-open-mpic-layer.arn]
@@ -108,7 +108,7 @@ resource "aws_api_gateway_deployment" "deployment" {
 resource "aws_lambda_permission" "lambda_permission_api" {
   statement_id  = "AllowOpenMPICAPIInvoke"
   action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.lambda_controller.function_name
+  function_name = aws_lambda_function.lambda_mpic_coordinator.function_name
   principal     = "apigateway.amazonaws.com"
 
   # The /* part allows invocation from any stage, method and resource path
