+++ open-tofu/validationwithcaa.tf
@@ -39,7 +39,7 @@ resource "aws_api_gateway_integration" "validationwithcaa_integration" {
 
   type = "AWS_PROXY"
 
-  uri = aws_lambda_function.lambda_controller.invoke_arn
+  uri = aws_lambda_function.lambda_mpic_coordinator.invoke_arn
