+++ src/aws_lambda_python/lambda_mpic_coordinator/mpic_coordinator_lambda_function.py
